+++ plugins/csv_collect.py
@@ -1,13 +1,27 @@
 import csv
 import timeit
 
-class csv_collect(object):
+from yapsy.IPlugin import IPlugin
+
+class PluginCSVCollect(IPlugin):
 	def __init__(self, file_name="collect.csv", delim = ","):
 		self.file_name = file_name
 		self.start_time = timeit.default_timer()
 		self.delim = delim
 
+	def activate(self, args):
+		if len(args) > 0:
+			self.file_name = args[0]
+		print "Will export CSV to:", self.file_name
 		open(self.file_name, 'w').close()
+		return True
+		
+	def deactivate(self):
+		#TODO: flush?
+		return
+
+	def show_help(self):
+		print "Optional argument: [filename] (default: collect.csv)"
 
 	def __call__(self, sample):
 		t = timeit.default_timer() - self.start_time
@@ -30,10 +44,3 @@ class csv_collect(object):
 		row += '\n'
 		with open(self.file_name, 'a') as f:
 			f.write(row)
-
-
-
-
-			
-
-