+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 Features:
   - Stream data over TCP (OpenViBE telnet reader format) or OSC
   - 16 channels support (daisy module)
-  - script: test sampling rate
+  - test sampling rate
+  - plugin system
+  - several different callback functions
 
 Bugfixes:
   - scale factor